requirements.txt
pyboy>=2.0.0
stable-baselines3>=2.0.0
gymnasium>=0.29.0
torch>=2.0.0
numpy>=1.24.0
python-dotenv
tensorboard
